gossip: move fetch functions
