+++ plugins/gossip/handler.go
@@ -1,7 +1,6 @@
 package gossip
 
 import (
-	"bytes"
 	"context"
 	"sync"
 	"time"
@@ -122,68 +121,6 @@ func (g *handler) HandleConnect(ctx context.Context, e muxrpc.Endpoint) {
 	<-ctx.Done()
 	g.Info.Log("msg", "fetchHops done", "n", hops.Count())
 
-}
-
-func (h *handler) fetchAllLib(ctx context.Context, e muxrpc.Endpoint, lst []librarian.Addr) error {
-	var refs = graph.NewFeedSet(len(lst))
-	for _, addr := range lst {
-		err := refs.AddAddr(addr)
-		if err != nil {
-			return err
-		}
-	}
-	return h.fetchAll(ctx, e, refs)
-}
-
-func (h *handler) fetchAllMinus(ctx context.Context, e muxrpc.Endpoint, fs graph.FeedSet, got []librarian.Addr) error {
-	lst, err := fs.List()
-	if err != nil {
-		return err
-	}
-	var refs = graph.NewFeedSet(len(lst))
-	for _, ref := range lst {
-		if !isIn(got, ref) {
-			err := refs.AddRef(ref)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return h.fetchAll(ctx, e, refs)
-}
-
-func (h *handler) fetchAll(ctx context.Context, e muxrpc.Endpoint, fs graph.FeedSet) error {
-	// we don't just want them all parallel right nw
-	// this kind of concurrency is way to harsh on the runtime
-	// we need some kind of FeedManager, similar to Blobs
-	// which we can ask for which feeds aren't in transit,
-	// due for a (probabilistic) update
-	// and manage live feeds more granularly across open connections
-
-	lst, err := fs.List()
-	if err != nil {
-		return err
-	}
-	for _, r := range lst {
-		err := h.fetchFeed(ctx, r, e)
-		if muxrpc.IsSinkClosed(err) {
-			return err
-		} else if err != nil {
-			// assuming forked feed for instance
-			h.Info.Log("msg", "fetchFeed stored failed", "err", err)
-		}
-		// }(r, &wg)
-	}
-	return nil
-	// wg.Wait()
-}
-func isIn(list []librarian.Addr, a *ssb.FeedRef) bool {
-	for _, el := range list {
-		if bytes.Compare([]byte(el), a.ID) == 0 {
-			return true
-		}
-	}
-	return false
 }
 
 func (g *handler) check(err error) {
